Add a confirm box when deleting components. #35

+++ src/manager/components/CommentItem/index.js
@@ -1,4 +1,4 @@
-/* eslint jsx-a11y/href-no-hash:0 */
+/* eslint jsx-a11y/href-no-hash:0, no-undef:0, no-alert:0 */
 
 import React, { Component } from 'react';
 import moment from 'moment';
@@ -27,7 +27,10 @@ export default class CommentItem extends Component {
   }
 
   deleteComment() {
-    this.props.deleteComment();
+    const confirmDelete = confirm('Are you sure you want to delete this comment?');
+    if (confirmDelete === true) {
+      this.props.deleteComment();
+    }
   }
 
   renderDelete() {
@@ -35,7 +38,7 @@ export default class CommentItem extends Component {
       <a
         href="#"
         style={style.commentDelete}
-        onClick={this.props.deleteComment}
+        onClick={this.deleteComment}
       >
         delete
       </a>
